+++ api/build.gradle.kts
@@ -5,10 +5,14 @@ plugins {
 group = "net.skinsrestorer"
 version = "14.1.6-SNAPSHOT"
 
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    implementation(project(":shared"))
+    compileOnly("org.projectlombok:lombok:1.18.22")
+    annotationProcessor("org.projectlombok:lombok:1.18.22")
+
+    compileOnly("org.spigotmc:spigot-api:1.17.1-R0.1-SNAPSHOT")
+    compileOnly("com.mojang:authlib:1.11")
+    compileOnly("net.md-5:bungeecord-api:1.17-R0.1-SNAPSHOT")
+    compileOnly("net.md-5:bungeecord-proxy:1.17-R0.1-SNAPSHOT")
+    compileOnly("com.velocitypowered:velocity-api:3.0.1")
+    compileOnly("org.spongepowered:spongeapi:7.3.0")
 }
