--- api/src/main/java/net/skinsrestorer/api/bukkit/events/SkinApplyBukkitEvent.java
@@ -1,56 +1,0 @@
-/*
- * SkinsRestorer
- *
- * Copyright (C) 2021 SkinsRestorer
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-3.0.html>.
- */
-package net.skinsrestorer.api.bukkit.events;
-
-import lombok.Getter;
-import lombok.Setter;
-import net.skinsrestorer.api.property.IProperty;
-import org.bukkit.entity.Player;
-import org.bukkit.event.Cancellable;
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
-import org.jetbrains.annotations.NotNull;
-
-@Getter
-public class SkinApplyBukkitEvent extends Event implements Cancellable {
-    public static final HandlerList HANDLERS = new HandlerList();
-    private final Player who;
-    @Setter
-    private IProperty property;
-    @Setter
-    private boolean isCancelled = false;
-
-    public SkinApplyBukkitEvent(@NotNull Player who, IProperty property) {
-        super(true);
-        this.property = property;
-        this.who = who;
-    }
-
-    @SuppressWarnings("unused")
-    public static HandlerList getHandlerList() {
-        return HANDLERS;
-    }
-
-    @NotNull
-    @Override
-    public HandlerList getHandlers() {
-        return HANDLERS;
-    }
-}
