+++ bukkit/build.gradle.kts
@@ -10,9 +10,17 @@ dependencies {
     compileOnly("org.projectlombok:lombok:1.18.22")
     annotationProcessor("org.projectlombok:lombok:1.18.22")
 
+    compileOnly("org.spigotmc:spigot-api:1.17.1-R0.1-SNAPSHOT")
+
     shadow("io.papermc:paperlib:1.0.6")
     shadow("org.bstats:bstats-bukkit:2.2.1")
 
     shadow("co.aikar:acf-paper:0.5.0-SNAPSHOT")
     shadow("com.github.cryptomorin:XSeries:8.4.0")
+
+    compileOnly("com.viaversion:viabackwards-common:4.0.1")
+    compileOnly("com.viaversion:viaversion:4.0.0")
+
+    implementation(project(":api"))
+    implementation(project(":shared"))
 }
