--- shared/src/main/java/li/cock/ie/reflect/DuckHandler.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2019 Ilya Egorov <ie9@null.net>. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package li.cock.ie.reflect;
-
-public class DuckHandler implements IDuckHandler {
-    private final boolean debug;
-    private Throwable ex = null;
-
-    public DuckHandler(boolean debug) {
-        this.debug = debug;
-    }
-
-    @Override
-    public void process(Throwable ex) {
-        if (debug) {
-            ex.printStackTrace();
-        }
-
-        this.ex = ex;
-    }
-
-    @Override
-    public void reset() {
-        this.ex = null;
-    }
-
-    @Override
-    public boolean check() {
-        boolean isSuccess = (ex == null);
-        reset();
-
-        return isSuccess;
-    }
-}
