--- shared/src/main/java/net/skinsrestorer/shared/exception/FieldNotFoundException.java
@@ -1,28 +1,0 @@
-/*
- * SkinsRestorer
- *
- * Copyright (C) 2021 SkinsRestorer
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-3.0.html>.
- */
-package net.skinsrestorer.shared.exception;
-
-public class FieldNotFoundException extends ReflectiveOperationException {
-    private static final long serialVersionUID = -1157594181223223976L;
-
-    public FieldNotFoundException(String message) {
-        super(message);
-    }
-}
