+++ shared/src/main/java/net/skinsrestorer/shared/utils/log/console/BungeeConsoleImpl.java
@@ -22,7 +22,7 @@ package net.skinsrestorer.shared.utils.log.console;
 import lombok.RequiredArgsConstructor;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.chat.TextComponent;
-import net.skinsrestorer.shared.interfaces.ISRConsole;
+import net.skinsrestorer.api.interfaces.ISRConsole;
 
 @RequiredArgsConstructor
 public class BungeeConsoleImpl implements ISRConsole {
